+++ src/main/java/com/gloddy/server/apply/service/ApplyService.java
@@ -6,8 +6,7 @@ import com.gloddy.server.apply.entity.Apply;
 import com.gloddy.server.apply.entity.vo.Status;
 import com.gloddy.server.apply.repository.ApplyJpaRepository;
 import com.gloddy.server.auth.entity.User;
-import com.gloddy.server.auth.repository.UserRepository;
-import com.gloddy.server.group.dto.GroupResponse;
+import com.gloddy.server.user.repository.UserRepository;
 import com.gloddy.server.group.entity.Group;
 import com.gloddy.server.group.repository.GroupJpaRepository;
 import lombok.RequiredArgsConstructor;
