+++ src/main/java/com/gloddy/server/auth/security/UserDetailsServiceImpl.java
@@ -3,7 +3,7 @@ package com.gloddy.server.auth.security;
 
 import com.gloddy.server.auth.entity.User;
 import com.gloddy.server.auth.jwt.JwtUserAdapter;
-import com.gloddy.server.auth.repository.UserRepository;
+import com.gloddy.server.user.repository.UserRepository;
 import com.gloddy.server.core.error.handler.errorCode.ErrorCode;
 import com.gloddy.server.core.error.handler.exception.UserBusinessException;
 import lombok.RequiredArgsConstructor;
